graphs tools : refactor
