+++ module/move/graphs_tools/src/search.rs
@@ -38,6 +38,7 @@ mod private
     Visit : FnMut( &'a Graph::Node ),
     Method : super::Method,
   {
+    /// Search traversing each node in an order specified by method.
     pub fn search( self, graph : &'a Graph )
     {
       graph.search( self )
@@ -104,91 +105,20 @@ mod private
       Self : Sized;
   }
 
-  /// Depth-first search strategy.
-  #[ derive( Debug, Default ) ]
-  pub struct Dfs;
-
-  impl Method for Dfs
-  {
-    type ExtraOptions = ();
-
-    /// Perform depth-first search on a graph.
-    fn _search< 'a, Graph, Visit >
-    (
-      graph : &'a Graph,
-      mut o : Options< 'a, Self, Graph, Visit >,
-    )
-    where
-      Visit : FnMut( &'a Graph::Node ),
-      Graph : ForGraphDirected< 'a > + ?Sized,
-    {
-      let mut visited = collection_tools::HashSet::new();
-      let mut stack = collection_tools::Vec::new();
-      stack.push( o.start_id );
-
-      while let Some( node_id ) = stack.pop()
-      {
-        let node = graph.node_ref( node_id );
-        if visited.insert( node_id )
-        {
-          ( o.visit )( node );
-          for child_id in graph.node_out_nodes( node_id )
-          {
-            stack.push( child_id );
-          }
-        }
-      }
-    }
-  }
-
-  /// Breadth-first search strategy.
-  #[ derive( Debug, Default ) ]
-  pub struct Bfs;
-
-  impl Method for Bfs
-  {
-    type ExtraOptions = ();
-
-    /// Perform breadth-first search on a graph.
-    fn _search< 'a, Graph, Visit >
-    (
-      graph : &'a Graph,
-      mut o : Options< 'a, Self, Graph, Visit >,
-    )
-    where
-      Visit : FnMut( &'a Graph::Node ),
-      Graph : ForGraphDirected< 'a > + ?Sized,
-    {
-      let mut visited = collection_tools::HashSet::new();
-      let mut queue = collection_tools::VecDeque::new();
-      queue.push_back( o.start_id );
-
-      while let Some( node_id ) = queue.pop_front()
-      {
-        let node = graph.node_ref( node_id );
-        if visited.insert( node_id )
-        {
-          ( o.visit )( node );
-          for child_id in graph.node_out_nodes( node_id )
-          {
-            queue.push_back( child_id );
-          }
-        }
-      }
-    }
-  }
-
 }
 
 crate::mod_interface!
 {
+  layer
+  {
+    dfs,
+    bfs,
+  };
   own use
   {
     options,
     Method,
     Options,
     ForGraphDirected,
-    Dfs,
-    Bfs,
   };
 }
