+++ module/move/graphs_tools/tests/inc/graph.rs
@@ -1,0 +1,3 @@
+use super::*;
+
+pub mod map_of_nodes;
