--- module/move/graphs_tools/tests/inc/search.rs
@@ -1,4 +1,0 @@
-use super::*;
-
-mod dfs;
-mod bfs;
