+++ module/move/graphs_tools/tests/inc/search_test.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+mod dfs_test;
+mod bfs_test;
